+++ client/src/components/Adopt/Adopt.js
@@ -2,63 +2,21 @@ import React, { useEffect, useState } from 'react'
 import Card from './Card.js'
 import banner from '../../assets/banneradopt.png'
 import steps from '../../assets/adoptSteps.png'
-import {  useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import style from './Adopt.module.css'
-import { filtersConfig, petsFiltered } from '../../redux/actions'
-import { getAllPets } from '../../redux/actions/index.js';
-import FilterByType from './Filters/FilterByType.js';
-import FilterByFoundation from './Filters/FilterByFoundation.js';
-import FilterByGender from './Filters/FilterByGender.js';
-import SearchBar from './Filters/SearchBar.js';
-import FilterByLocation from './Filters/FilterByLocation.js';
 import Paginate from './Filters/Paginate.js';
+import Filters from './Filters/Filters.js';
+import Loading from '../Loading/Loading'
 
 
 export default function Adopt(){
     
-    const dispatch = useDispatch()
     const {filtered,pages} = useSelector(state => state.petsFiltered)
     
-    useEffect(()=>{
-        dispatch(getAllPets())
-    },[dispatch])
-
-    const arrAllPets = useSelector(state=> state.allPets )
-    const filters = useSelector(state=> state.filtersConfig )
-
-    //In case browser reload
-    useEffect(()=>{
-        return function clean(){
-            dispatch(filtersConfig({type:null,foundation:null,gender:null }))}
-    },[])
-
-
-    useEffect(()=>{
-        let filteringPets = arrAllPets
-
-        if(filters.type ){
-            filteringPets = filteringPets.filter(pet => pet.type === filters.type)
-        }
-        if(filters.foundation){
-            filteringPets = filteringPets.filter(pet => pet.foundation.name === filters.foundation)
-        }
-        if(filters.gender){
-            filteringPets = filteringPets.filter(pet => pet.gender === filters.gender)
-        }
-        if(filters.city){
-            filteringPets = filteringPets.filter(pet => pet.foundation.name === filters.city)
-        }
-        if(filters.name){
-            filteringPets = filteringPets.filter(pet => (pet.name).toLowerCase().includes((filters.name).toLowerCase()))
-        }
-        if(!filteringPets.length && arrAllPets.length && !filters.name) alert("No hay animales en adopcion con esas caracteristicas")
-        
-        dispatch(petsFiltered(filteringPets)) 
-        
-    },[filters,arrAllPets])
+    let renderPerPage = [] 
+    if(pages) renderPerPage = pages
+    else if (filtered) renderPerPage = filtered.slice(0,12)
     
-
-    let renderPerPage = pages || arrAllPets.slice(0,12)
     return (
         <>
             <div className={style.banner}>
@@ -89,24 +47,8 @@ export default function Adopt(){
             </div>
             <h2 className={style.innerTittles} >Conoce nuestras Huellas</h2>
             <div className={style.allAnimals}>
-                <div className={style.allAnimals__filters}>
-                        <FilterByType/>
-                        <SearchBar/>
-                        <div className={style.containerToggle}>
-                            <div className={style.filterIcon}>
-                                <i className="fa-solid fa-filter"></i>
-                                <i className="fa-solid fa-bars"></i>
-                            </div>
-                            <div className={style.containerFiltersBy}>
-                                <div className={style.filtersBy}>
-                                    <FilterByFoundation/>
-                                    <FilterByGender/>
-                                    <FilterByLocation/>
-                                </div>
-                            </div>
-                    </div>
-                </div>
-                </div>
+                <Filters/>
+            </div>
             <div className={style.cardContainer}>
                 <Paginate/>
                 {renderPerPage ?
@@ -119,8 +61,7 @@ export default function Adopt(){
                         />
                     ))
                     :
-                    //Add loading component
-                    <h3>Cargando..</h3>
+                    <Loading/>
                 }
             </div>
         </>
